sdk: refactor: split services epics
...into pathfinding, monitor and udc epics, since the generic epic
module was already too big.
